server/requirements/requirements-base.txt
Django==3.1.8
djangorestframework==3.12.4
django-rest-passwordreset==1.1.0
django-ckeditor==6.0.0
psycopg2-binary==2.8.6